# be/cmd/server/docs/swagger.yaml
definitions:
...
  dto.CreateCategoryRequest:
    properties:
      categoryName:
        type: string
    required:
    - categoryName
    type: object
  dto.CreateDepartmentRequest:
    properties:
      departmentName:
        type: string
      locationId:
        type: integer
    required:
    - departmentName
    - locationId
    type: object
  dto.CreateLocationRequest:
    properties:
      locationName:
        type: string
    required:
    - locationName
    type: object
...
info:
  contact: {}
paths:
...
  /api/categories:
    get:
      consumes:
      - application/json
      description: Get all categories
      produces:
      - application/json
      responses: {}
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create categories
      parameters:
      - description: Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses: {}
      summary: Create categories
      tags:
      - Categories
  /api/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete category via id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete category
      tags:
      - Categories
  /api/departments:
    get:
      consumes:
      - application/json
      description: Get all departments
      produces:
      - application/json
      responses: {}
      summary: Get all departments
      tags:
      - Departments
    post:
      consumes:
      - application/json
      description: Create departments
      parameters:
      - description: Data
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDepartmentRequest'
      produces:
      - application/json
      responses: {}
      summary: Create departments
      tags:
      - Departments
  /api/departments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete department via id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete department
      tags:
      - Departments
  /api/locations:
    get:
      consumes:
      - application/json
      description: Get all location
      produces:
      - application/json
      responses: {}
      summary: Get all location
      tags:
      - Locations
    post:
      consumes:
      - application/json
      description: Create location
      parameters:
      - description: Data
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLocationRequest'
      produces:
      - application/json
      responses: {}
      summary: Create Location
      tags:
      - Locations
  /api/locations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete location via id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete location
      tags:
      - Locations
...
swagger: "2.0"